refactor: flush TODOs: clean up ELike and EDumpItem
