+++ packages/akashic-cli-serve/src/client/store/DevtoolUiStore.ts
@@ -1,5 +1,5 @@
 import { action, observable, ObservableMap } from "mobx";
-import { EDumpItem } from "../akashic/EDumpItem";
+import { EDumpItem } from "../common/types/EDumpItem";
 import { storage } from "./storage";
 
 export class DevtoolUiStore {
